ROUTES:dashboard
- Changed dashboard route to be /dashboard
- Removed unused routes file
