+++ controllers/dashboard-routes.js
@@ -1,18 +1,31 @@
-const withAuth = require("../utils/auth");
 const router = require("express").Router();
-const db = require("../models");
+const employee = require("../models/employee");
+const role = require("../models/role");
+const department = require("../models/department");
 
-router.get("/", async (req, res) => {
-  try {
-    const employees = await db.employee.findAll({
-      include: [db.role], // Include the 'role' model in the query
+router.get('/dashboard', (req, res) => {
+  employee.findAll({
+    include: [{ model: role, include: [department] }]
+  })
+    .then((employees) => {
+      const processedData = employees.map((employee) => {
+        return {
+          id: employee.id,
+          first_name: employee.first_name,
+          last_name: employee.last_name,
+          role: employee.role.title,
+          salary: employee.role.salary,
+          department_id: employee.role.department.id,
+          department_name: employee.role.department.department_name
+        };
+      });
+
+      res.render('homepage', { processedData: JSON.stringify(processedData) });
+    })
+    .catch((error) => {
+      console.error('Error:', error);
+      res.status(500).send('An error occurred');
     });
-
-    res.render("dashboard", { employees });
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ error: "Internal server error" });
-  }
 });
 
 module.exports = router;
