+++ controllers/index.js
@@ -1,10 +1,8 @@
 const router = require("express").Router();
 
 const apiRoutes = require("./api");
-const dashboardRoutes = require("./dashboard-routes.js");
-const homeRoutes = require("./homeRoutes.js");
+const dashboardRoutes = require("./homeRoutes.js");
 
-router.use("/", homeRoutes);
 router.use("/dashboard", dashboardRoutes);
 router.use("/api", apiRoutes);
